fix warnings
Fix a couple warnings so we can use meson warning level 3.  Instances of
the following warnings are found and resolved:

  Wreorder
  Wsign-compare

+++ include/HwmonTempSensor.hpp
@@ -24,7 +24,7 @@ class HwmonTempSensor : public Sensor
     boost::asio::deadline_timer waitTimer;
     boost::asio::streambuf readBuf;
     std::string path;
-    int errCount;
+    size_t errCount;
     void setupRead(void);
     void handleResponse(const boost::system::error_code& err);
     void checkThresholds(void) override;
